#1918 移动通信4G网优软件平台功能增强

+++ src/main/java/com/hgicreate/rno/web/rest/LteDtDataResource.java
@@ -44,30 +44,37 @@ public class LteDtDataResource {
 
     /**
      * 接收上传文件并存储为本地文件
+     *
      * @return 成功情况下返回 HTTP OK 状态，错误情况下返回 HTTP 4xx 状态。
      */
     @PostMapping("/upload-file")
     public ResponseEntity<?> uploadFile(FileUploadVM vm) {
 
         log.debug("模块名：" + vm.getModuleName());
-        log.debug("上传文件处理中...");
 
         try {
             // 获取文件名，并构建为本地文件路径
             String filename = vm.getFile().getOriginalFilename();
-            log.debug("filename = {}", filename);
+            log.debug("上传的文件名：{}", filename);
+
+            // 如果目录不存在则创建目录
+            File fileDirectory = new File(directory);
+            if (!fileDirectory.exists() && !fileDirectory.mkdirs()) {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
 
             // 以随机的 UUID 为文件名存储在本地
             filename = UUID.randomUUID().toString();
             String filepath = Paths.get(directory, filename).toString();
+
+            log.debug("存储的文件名：{}", filename);
 
             // 保存文件到本地
             BufferedOutputStream stream =
                     new BufferedOutputStream(new FileOutputStream(new File(filepath)));
             stream.write(vm.getFile().getBytes());
             stream.close();
-        }
-        catch (Exception e) {
+        } catch (Exception e) {
             System.out.println(e.getMessage());
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
